Refactored to use mage directories
