+++ go.mod
@@ -1,8 +1,8 @@
 module github.com/henryhall897/golang-todo-app
 
-go 1.23
+go 1.24
 
-toolchain go1.23.5
+toolchain go1.24.0
 
 require (
 	github.com/golang-migrate/migrate/v4 v4.18.1
@@ -50,9 +50,11 @@ require (
 	go.uber.org/atomic v1.11.0 // indirect
 	go.uber.org/multierr v1.11.0 // indirect
 	golang.org/x/crypto v0.31.0 // indirect
-	golang.org/x/sync v0.10.0 // indirect
-	golang.org/x/sys v0.28.0 // indirect
+	golang.org/x/sync v0.11.0 // indirect
+	golang.org/x/sys v0.30.0 // indirect
 	golang.org/x/text v0.21.0 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
+
+replace github.com/henryhall897/golang-todo-app/mage => ./magefiles
